更新矿机存储逻辑

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/eth_hub/entity/MinerDashboard.java
@@ -27,7 +27,7 @@ public class MinerDashboard {
         public static class CurrentStatisticsDTO {
             private Integer time;
             private Integer lastSeen;
-            private Long reportedHashrate;
+            private Double reportedHashrate;
             private Double currentHashrate;
             private Integer validShares;
             private Integer invalidShares;
